requirements.txt
streamlit
scikit-learn
pandas
numpy
plotly
Sastrawi
python-Levenshtein
joblib
scipy
